FEAT: Build for F-droid.
