+++ VirtualApp/app/src/fdroid/java/io/virtualapp/delegate/MyVirtualInitializer.java
@@ -1,0 +1,15 @@
+package io.virtualapp.delegate;
+
+import android.app.Application;
+
+import com.lody.virtual.client.core.VirtualCore;
+
+/**
+ * @author weishu
+ * @date 2019/2/25.
+ */
+public class MyVirtualInitializer extends BaseVirtualInitializer {
+    public MyVirtualInitializer(Application application, VirtualCore core) {
+        super(application, core);
+    }
+}
