+++ VirtualApp/app/src/main/java/io/virtualapp/VCommends.java
@@ -13,5 +13,4 @@ public class VCommends {
 	public static final String EXTRA_APP_INFO_LIST = "va.extra.APP_INFO_LIST";
 
 	public static final String TAG_ASK_INSTALL_GMS = "va.extra.ASK_INSTALL_GMS";
-
 }
