+++ VirtualApp/app/src/main/java/io/virtualapp/XApp.java
@@ -2,23 +2,13 @@ package io.virtualapp;
 
 import android.app.Application;
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.os.Build;
 
-import com.crashlytics.android.Crashlytics;
 import com.lody.virtual.client.NativeEngine;
 import com.lody.virtual.client.core.VirtualCore;
 import com.lody.virtual.client.stub.VASettings;
-import com.lody.virtual.os.VEnvironment;
 
-import io.fabric.sdk.android.Fabric;
-import io.virtualapp.delegate.MyAppRequestListener;
-import io.virtualapp.delegate.MyComponentDelegate;
-import io.virtualapp.delegate.MyCrashHandler;
-import io.virtualapp.delegate.MyPhoneInfoDelegate;
-import io.virtualapp.delegate.MyTaskDescDelegate;
-import jonathanfinerty.once.Once;
-import me.weishu.exposed.LogcatService;
+import io.virtualapp.delegate.MyVirtualInitializer;
 
 /**
  * @author Lody
@@ -30,7 +20,6 @@ public class XApp extends Application {
     public static final String XPOSED_INSTALLER_PACKAGE = "de.robv.android.xposed.installer";
 
     private static XApp gApp;
-    private SharedPreferences mPreferences;
 
     public static XApp getApp() {
         return gApp;
@@ -43,7 +32,6 @@ public class XApp extends Application {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
             NativeEngine.disableJit(Build.VERSION.SDK_INT);
         }
-        mPreferences = base.getSharedPreferences("va", Context.MODE_MULTI_PROCESS);
         VASettings.ENABLE_IO_REDIRECT = true;
         VASettings.ENABLE_INNER_SHORTCUT = false;
         try {
@@ -57,47 +45,7 @@ public class XApp extends Application {
     public void onCreate() {
         super.onCreate();
         VirtualCore virtualCore = VirtualCore.get();
-        virtualCore.initialize(new VirtualCore.VirtualInitializer() {
-
-            @Override
-            public void onMainProcess() {
-                Once.initialise(XApp.this);
-                Fabric.with(XApp.this, new Crashlytics());
-            }
-
-            @Override
-            public void onVirtualProcess() {
-                Fabric.with(XApp.this, new Crashlytics());
-
-                //listener components
-                virtualCore.setComponentDelegate(new MyComponentDelegate());
-                //fake phone imei,macAddress,BluetoothAddress
-                virtualCore.setPhoneInfoDelegate(new MyPhoneInfoDelegate());
-                //fake task description's icon and title
-                virtualCore.setTaskDescriptionDelegate(new MyTaskDescDelegate());
-                virtualCore.setCrashHandler(new MyCrashHandler());
-
-                // ensure the logcat service alive when every virtual process start.
-                LogcatService.start(XApp.this, VEnvironment.getDataUserPackageDirectory(0, XPOSED_INSTALLER_PACKAGE));
-            }
-
-            @Override
-            public void onServerProcess() {
-                virtualCore.setAppRequestListener(new MyAppRequestListener(XApp.this));
-                virtualCore.addVisibleOutsidePackage("com.tencent.mobileqq");
-                virtualCore.addVisibleOutsidePackage("com.tencent.mobileqqi");
-                virtualCore.addVisibleOutsidePackage("com.tencent.minihd.qq");
-                virtualCore.addVisibleOutsidePackage("com.tencent.qqlite");
-                virtualCore.addVisibleOutsidePackage("com.facebook.katana");
-                virtualCore.addVisibleOutsidePackage("com.whatsapp");
-                virtualCore.addVisibleOutsidePackage("com.tencent.mm");
-                virtualCore.addVisibleOutsidePackage("com.immomo.momo");
-            }
-        });
-    }
-
-    public static SharedPreferences getPreferences() {
-        return getApp().mPreferences;
+        virtualCore.initialize(new MyVirtualInitializer(this, virtualCore));
     }
 
 }
